Fix C and C++ functions to actualy take 8-byte integers

+++ cbits/sort_cplusplus.cpp
@@ -1,7 +1,8 @@
 #include <bits/stdc++.h>
+#include <cstdint>
 
 extern "C" {
-    void cplusplus_sort(int* xs, int len) {
+    void cplusplus_sort(int64_t* xs, int len) {
         std::sort(xs, xs + len);
     }
 }
